fix: using global atomic for peak tracking for better accuracy
This change drastically increases accuracy, the previous approach would
over report allocations by 2 orders of maginitude in some cases (low
hundreds of MBs to dozens of GBs). The issue is that an individual
thread can not accuratly account for all allocations/deallocations for
objects that are sent across thread boundaries.