feat: new global allocator
This introduces a global allocator to track peak memory usage, with much
better accuracy than the thread based allocator tracker. The downside of
the global version is that it has a single atomic used for all threads,
which may be a source of contention. The advantage is that it is
accurate and has much simpler code.