feat: drastically improves performance by using libc::backtrace
Backtrace::new_unresolved has a lot of overhead and it is not suitable
for collection traces on every allocation. The slow down is extreme and
make programs unusable.

This pathc changes the implementation to use libc::backtrace directly,
it only collects the instruction pointers and delays resolving the
pointers to function names to the formatter. Writing the flamegraph is
still a considerably expensive operation, but it can be delegated to a
background thread leaving the allocator unblocked so the application can
continue working.